Move gc-specific `Store` methods to `store/gc.rs`
Keeps imports a bit more organized and additionally cuts down on the
quantity of #[cfg] needed.
